requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# AI/ML
google-genai==1.0.0

# Data Storage
redis==5.0.1

# HTTP Requests
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Optional: Enhanced logging and monitoring
python-json-logger==2.0.7